% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{DV_plot}
\alias{DV_plot}
\title{Plot 3D scatter plot in two conditions}
\usage{
DV_plot(df, targetgene = NULL, ptSize = 3, lwd = 5, dlwd = 7)
}
\arguments{
\item{df}{Resultant data.frame after DV_splinefit analysis}

\item{targetgene}{An integer value. Defines the minimum reads required for a cell to be included in the analysis.}

\item{ptSize}{An integer value. Defines point size for dots}

\item{lwd}{An integer value. Defines line width for spline.}

\item{dlwd}{An integer value. Defines Line width for target gene distance.}
}
\value{
3D plotly scatter plot
}
\description{
Plots 3D gene statistic scatter plot of target genes in twop conditions
}
\examples{
# example code
## Load Data
load(system.file("extdata", "WT_count.rda", package="SplineDV")) # WT Sample
load(system.file("extdata", "KO_count.rda", package="SplineDV")) # KO Sample
DV_res <- DV_splinefit(X=KO_count, Y=WT_count)
fig <- DV_plot(DV_res)
print(fig)
}
\author{
Shreyan Gupta <xenon8778@tamu.edu>
}
