% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{DV_splinefit}
\alias{DV_splinefit}
\title{Spline-DV}
\usage{
DV_splinefit(X = x, Y = y, nfeatures = 500, ncells = 15, mt.perc = 15)
}
\arguments{
\item{X}{Count matrix or SeuratObject (Test sample)}

\item{Y}{Count matrix or SeuratObject (Control sample)}

\item{nfeatures}{An integer value. Defines the minimum features (genes) required for a cell to be included in the analysis.}

\item{ncells}{An integer value. Defines the minimum cells required for a gene to be included in the analysis.}

\item{mt.perc}{A double value. Defines the minimum percent mitochondrial genes expression required for a cell to be excluded from the analysis.}
}
\value{
A DataFrame with DV Statistics
}
\description{
Differential Variability analysis
}
\examples{
# example code
# Load Data
exprMatrix_WT = read.csv('https://github.com/Xenon8778/SplineDV/raw/main/data/WTdata.csv', row.names = 1) # WT Sample
exprMatrix_KO = read.csv('https://github.com/Xenon8778/SplineDV/raw/main/data/KOdata.csv', row.names = 1) # KO Sample
DV_res = DV_splinefit(X = exprMatrix_KO, Y = exprMatrix_WT)
}
\author{
Shreyan Gupta <xenon8778@tamu.edu>
}
