% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{HVG_QC}
\alias{HVG_QC}
\title{Quality control}
\usage{
HVG_QC(X = x, nfeatures = 500, ncells = 15, mt.perc = 15)
}
\value{
QC filtered SeuratObject
}
\description{
QC filter scRNA-seq expression data
}
\author{
Shreyan Gupta <xenon8778@tamu.edu>
}
