% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{HVG_splinefit}
\alias{HVG_splinefit}
\title{Spline-HVG}
\usage{
HVG_splinefit(
  X = x,
  QC = T,
  nfeatures = 500,
  ncells = 15,
  mt.perc = 15,
  degf = 15,
  spar = 0.75,
  nHVGs = 2000,
  show.spline = FALSE,
  use.ndist = T
)
}
\description{
Compute Highly Variable Genes
}
\author{
Shreyan Gupta <xenon8778@tamu.edu>
Spline HVG function
This function computed highly variable genes from an scRNAseq count matrix.
}
