% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{HVGPlot}
\alias{HVGPlot}
\title{Plot HVG 3D scatter plot}
\usage{
HVGPlot(df, targetgene = NULL, ptSize = 3, lwd = 5, dlwd = 7)
}
\arguments{
\item{df}{Resultant data.frame after splineHVG analysis}

\item{targetgene}{An integer value. Defines the minimum reads required for a cell to be included in the analysis.}

\item{ptSize}{An integer value. Defines point size for dots}

\item{lwd}{An integer value. Defines line width for spline.}

\item{dlwd}{An integer value. Defines Line width for target gene distance.}
}
\value{
3D plotly scatter plot
}
\description{
Plots 3D scatter plot with HVGs or target gene highlighted
}
\examples{
# example code
## Load Data
load(system.file("extdata", "WT_count.rda", package="SplineDV")) # WT Sample
hvgDF <- splineHVG(WT_count)
fig <- HVGPlot(hvgDF)
print(fig)
}
\author{
Shreyan Gupta <xenon8778@tamu.edu>
}
