% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{hvgQC}
\alias{hvgQC}
\title{Quality control}
\usage{
hvgQC(X, ncounts = 1000, ncells = 15, mtPerc = 15)
}
\arguments{
\item{X}{Count matrix or SingleCellExperiment}

\item{ncounts}{An integer value. Defines the minimum reads required for a cell to be included in the analysis.}

\item{ncells}{An integer value. Defines the minimum cells required for a gene to be included in the analysis.}

\item{mtPerc}{A double value. Defines the minimum percent mitochondrial genes expression required for a cell to be excluded from the analysis.}
}
\value{
QC filtered SingleCellExperiment
}
\description{
QC filter scRNA-seq expression data
}
\examples{
# example code
## Load Data
load(system.file("extdata", "WT_count.rda", package="SplineDV")) # WT Sample
adata = hvgQC(WT_count)
}
\author{
Shreyan Gupta <xenon8778@tamu.edu>
}
