% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{KOcount}
\alias{KOcount}
\title{Nkx2-1 KO Sample}
\format{
A Matrix object.
}
\source{
<https://doi.org/10.1152/ajplung.00337.2015>
}
\usage{
KOcount
}
\description{
This is only the KO sample.
This example data is borrowed from an
experimental Nkx2-1 Gene knockout scRNA-seq study
}
\details{
details for KO_count
}
\keyword{datasets}
