% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{splineDV}
\alias{splineDV}
\title{Spline-DV}
\usage{
splineDV(X, Y, ncounts = 500, ncells = 15, mtPerc = 15, detailed = FALSE)
}
\arguments{
\item{X}{Count matrix or SingleCellExperiment (Test sample)}

\item{Y}{Count matrix or SingleCellExperiment (Control sample)}

\item{ncounts}{An integer value. Defines the minimum reads required for a cell to be included in the analysis.}

\item{ncells}{An integer value. Defines the minimum cells required for a gene to be included in the analysis.}

\item{mtPerc}{A double value. Defines the minimum percent mitochondrial genes expression required for a cell to be excluded from the analysis.}

\item{detailed}{A boolean value. Defines whether to add individual splineHVG data.frame to the output.}
}
\value{
A data.frame with DV Statistics
}
\description{
Differential Variability analysis
}
\examples{
# example code
# Load Data
WTcount <- SplineDV::WTcount # WT Sample
KOcount <- SplineDV::KOcount # KO Sample
DVres <- splineDV(X=KOcount, Y=WTcount)
}
\author{
Shreyan Gupta <xenon8778@tamu.edu>
}
