% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{splineHVG}
\alias{splineHVG}
\title{Spline-HVG}
\usage{
splineHVG(
  X,
  QC = TRUE,
  ncounts = 500,
  ncells = 15,
  mtPerc = 15,
  degf = 15,
  spar = 0.75,
  nHVGs = 2000,
  use.ndist = TRUE
)
}
\arguments{
\item{X}{Count matrix or SingleCellExperiment}

\item{QC}{A Boolean value (TRUE/FALSE), if TRUE, a quality control is applied over the data.}

\item{ncounts}{An integer value. Defines the minimum reads required for a cell to be included in the analysis.}

\item{ncells}{An integer value. Defines the minimum cells required for a gene to be included in the analysis.}

\item{mtPerc}{A double value. Defines the minimum percent mitochondrial genes expression required for a cell to be excluded from the analysis.}

\item{degf}{An integer value. Degrees of freedom for the Spline.}

\item{spar}{A double value. Smoothing parameter for Spline}

\item{nHVGs}{An integer value. Number of top Highly Variable Genes (HVGs) to select.}

\item{use.ndist}{A Boolean value (TRUE/FALSE), if TRUE, computes the nearest point on spline by nearest-neighbor search (TRUE Recommended). Else, uses the position of the corresponding gene on the spline for distance computation.}
}
\value{
A data.frame with HVG selection Statistics
}
\description{
Compute Highly Variable Genes from an scRNAseq expression data.
}
\examples{
# example code
## Load Data
load(system.file("extdata", "WT_count.rda", package = "SplineDV")) # WT Sample
HVG_res <- splineHVG(WT_count, nHVGs = 100)
}
\author{
Shreyan Gupta <xenon8778@tamu.edu>
}
