% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{WT_count}
\alias{WT_count}
\title{Nkx2-1 WT Sample}
\format{
A Matrix object.
}
\source{
<https://doi.org/10.1152/ajplung.00337.2015>
}
\usage{
WT_count
}
\description{
This is only the WT sample.
This example data is borrowed from an
experimental Nkx2-1 Gene knockout scRNA-seq study
}
\details{
details for WT_count
}
\keyword{datasets}
