% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{DV_splinefit}
\alias{DV_splinefit}
\title{Spline-DV}
\usage{
DV_splinefit(
  X = x,
  Y = y,
  nfeatures = 500,
  ncells = 15,
  mt.perc = 15,
  nHVGs = 2000
)
}
\value{
A DataFrame with DV Statistics
}
\description{
Differential Variability analysis
}
\author{
Shreyan Gupta <xenon8778@tamu.edu>
}
