% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{DVPlot}
\alias{DVPlot}
\title{Plot 3D scatter plot in two conditions}
\usage{
DVPlot(df, targetgene = NULL, ptSize = 3, lwd = 5, dlwd = 7)
}
\arguments{
\item{df}{Resultant data.frame after splineDV analysis}

\item{targetgene}{An integer value. Defines the minimum reads required for a cell to be included in the analysis.}

\item{ptSize}{An integer value. Defines point size for dots}

\item{lwd}{An integer value. Defines line width for spline.}

\item{dlwd}{An integer value. Defines Line width for target gene distance.}
}
\value{
3D plotly scatter plot
}
\description{
Plots 3D gene statistic scatter plot of target genes in twop conditions
}
\examples{
# example code
## Load Data
WTcount <- SplineDV::WTcount # WT Sample
KOcount <- SplineDV::KOcount # KO Sample
DVres <- splineDV(X=KOcount, Y=WTcount)
fig <- DVPlot(DVres)
print(fig)
}
\author{
Shreyan Gupta <xenon8778@tamu.edu>
}
