% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{HVG_QC}
\alias{HVG_QC}
\title{Quality control}
\usage{
HVG_QC(X = x, ncounts = 1000, ncells = 15, mt.perc = 15)
}
\arguments{
\item{X}{Count matrix or SingleCellExperiment}

\item{ncounts}{An integer value. Defines the minimum reads required for a cell to be included in the analysis.}

\item{ncells}{An integer value. Defines the minimum cells required for a gene to be included in the analysis.}

\item{mt.perc}{A double value. Defines the minimum percent mitochondrial genes expression required for a cell to be excluded from the analysis.}
}
\value{
QC filtered SingleCellExperiment
}
\description{
QC filter scRNA-seq expression data
}
\examples{
# example code
## Load Data
WT_count <- get(data("WT_count", package = 'SplineDV')) # WT Sample
adata = HVG_QC(WT_count)
}
\author{
Shreyan Gupta <xenon8778@tamu.edu>
}
