% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{HVG_splinefit}
\alias{HVG_splinefit}
\title{Spline-HVG}
\usage{
HVG_splinefit(
  X = x,
  QC = TRUE,
  ncounts = 500,
  ncells = 15,
  mt.perc = 15,
  degf = 15,
  spar = 0.75,
  nHVGs = 2000,
  show.spline = FALSE,
  use.ndist = TRUE
)
}
\arguments{
\item{X}{Count matrix or SingleCellExperiment}

\item{QC}{A Boolean value (TRUE/FALSE), if TRUE, a quality control is applied over the data.}

\item{ncounts}{An integer value. Defines the minimum reads required for a cell to be included in the analysis.}

\item{ncells}{An integer value. Defines the minimum cells required for a gene to be included in the analysis.}

\item{mt.perc}{A double value. Defines the minimum percent mitochondrial genes expression required for a cell to be excluded from the analysis.}

\item{degf}{An integer value. Degrees of freedom for the Spline.}

\item{spar}{A double value. Smoothing parameter for Spline}

\item{nHVGs}{An integer value. Number of top Highly Variable Genes (HVGs) to select.}

\item{show.spline}{A Boolean value (TRUE/FALSE), if TRUE, shows the 3D Spline as Plotly figure.}

\item{use.ndist}{A Boolean value (TRUE/FALSE), if TRUE, computes the nearest point on spline by nearest-neighbor search (TRUE Recommended). Else, uses the position of the corresponding gene on the spline for distance computation.}
}
\value{
A DataFrame with HVG selection Statistics
}
\description{
Compute Highly Variable Genes from an scRNAseq expression data.
}
\examples{
# example code
## Load Data
exprMatrix = read.csv('https://github.com/Xenon8778/SplineDV/raw/main/data/WTdata.csv', row.names = 1) # WT Sample
HVG_res = HVG_splinefit(exprMatrix, nHVGs = 100)
}
\author{
Shreyan Gupta <xenon8778@tamu.edu>
}
